made beverage & food take in a float ionstead of int, added Experience class, add deletion methods to room & house, and made SystemManager have an arraylist of Experiences

+++ Beverage.java
@@ -1,6 +1,6 @@
 public class Beverage extends MenuItem {
 
-    public Beverage(String name, int cost) {
+    public Beverage(String name, float cost) {
         super(name, cost);
     }
